add helper function name_of_missing_import
Not quite sure if all these helper functions are pulling their weight

Being explicit about extracting a name out of the line seems to be a
step in the right direction, though.

Splitting and checking the result also works well, I think.
The previous solution (if x in line, line.split(x)) was a bit
repetitive).