create source file based on test file (not done)
Previously create_file.py created a test_file for every source file.
It tried to force you to create tests for each and every source file.
It even switched the editor to the source file automagically.

This is a bit too much since even people like Uncle Bob don't have write
unit tests for stuff like GUI.

One problem with the new approach is that we might guess the wrong name
for the source file. Example:
	You create test_some_class_some_scenario.py
	-> create_file creates some_class_some_scenario.py for you
	- but you only wanted to create some_class.py
This just shows that we need to tighten the TDD workflow and/or add a
config/plugin mechanism that allows the user to adapt save_file.py to
his workflow.